+++ help/outline.go
@@ -11,40 +11,18 @@ import (
 	"go.bbkane.com/warg/help/common"
 )
 
-func outlineFlagHelper(w io.Writer, color *gocolor.Color, flagName string, f cli.Flag, indent int) {
-	str := common.FmtFlagName(color, flagName)
-	if f.Alias != "" {
-		str = str + " , " + common.FmtFlagAlias(color, f.Alias)
-	}
-	fmt.Fprintln(w, common.LeftPad("# "+string(f.HelpShort), "  ", indent))
-	fmt.Fprintln(
-		w,
-		common.LeftPad(str, "  ", indent),
-	)
-}
-
 func outlineHelper(w io.Writer, color *gocolor.Color, sec cli.Section, indent int) {
 	// commands and command flags
 	for _, comName := range sec.Commands.SortedNames() {
-		com := sec.Commands[string(comName)]
-		fmt.Fprintln(w, common.LeftPad("# "+string(com.HelpShort), "  ", indent))
 		fmt.Fprintln(
 			w,
 			common.LeftPad(common.FmtCommandName(color, string(comName)), "  ", indent),
 		)
-		for _, flagName := range com.Flags.SortedNames() {
-			outlineFlagHelper(w, color, flagName, com.Flags[flagName], indent+1)
-		}
-
 	}
 
 	// sections
 	for _, k := range sec.Sections.SortedNames() {
 		childSec := sec.Sections[k]
-		fmt.Fprintln(
-			w,
-			common.LeftPad("# "+string(childSec.HelpShort), "  ", indent),
-		)
 		fmt.Fprintln(
 			w,
 			common.LeftPad(common.FmtSectionName(color, k), "  ", indent),